requirements.txt
certifi==2026.2.25
influxdb-client==1.50.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reactivex==4.1.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
